update scripts for gencast; add project setup file
