+++ .gitignore
@@ -1,0 +1,4 @@
+*.egg*
+__pycache__
+*.swp
+*vscode*
