+++ environment.yml
@@ -1,0 +1,18 @@
+name: graphcast
+channels:
+  - conda-forge
+dependencies:
+  - python==3.11
+  - dm-tree
+  - netcdf4
+  - xarray
+  - cartopy
+  - pygrib
+  - eccodes
+  - iris
+  - iris-grib
+  - jupyterlab
+  - boto3
+  - pip
+  - pip:
+    - https://github.com/deepmind/graphcast/archive/master.zip
