+++ oper/utils/nc2grib.py
@@ -20,7 +20,7 @@ class Netcdf2Grib:
             '10m_v_component_of_wind': [10, 'y_wind', 'm s**-1'],
             'mean_sea_level_pressure': [0, 'air_pressure_at_sea_level', 'Pa'],
             '2m_temperature': [2, 'air_temperature', 'K'],
-            'total_precipitation_6hr': [0, 'precipitation_amount', 'kg m**-2'],
+            'total_precipitation_12hr': [0, 'precipitation_amount', 'kg m**-2'],
             'total_precipitation': [0, 'precipitation_amount', 'kg m**-2'],
             'vertical_velocity': [None, 'lagrangian_tendency_of_air_pressure', 'Pa s**-1'],
             'specific_humidity': [None, 'specific_humidity', 'kg kg**-1'],
@@ -72,8 +72,8 @@ class Netcdf2Grib:
         forecasts['level'].attrs['long_name'] = 'pressure'
         forecasts['level'].attrs['units'] = 'Pa'
         forecasts['geopotential'] = forecasts['geopotential'] / 9.80665
-        forecasts['total_precipitation_6hr'] = forecasts['total_precipitation_6hr'] * 1000
-        forecasts['total_precipitation'] = forecasts['total_precipitation_6hr'].cumsum(axis=0)
+        forecasts['total_precipitation_12hr'] = forecasts['total_precipitation_12hr'] * 1000
+        forecasts['total_precipitation'] = forecasts['total_precipitation_12hr'].cumsum(axis=0)
 
         filename = os.path.join(outdir, f"forecast_to_grib2_{gefs_member}.nc")
         forecasts.to_netcdf(filename)
@@ -97,7 +97,7 @@ class Netcdf2Grib:
         for date in datevectors:
             print(f"Processing for time {date.strftime('%Y-%m-%d %H:00:00')}")
             hrs = int((date - forecast_starttime).total_seconds() // 3600)
-            outfile = os.path.join(outdir, f'graphcastge{gefs_member}.t{cycle:02d}z.pgrb2.0p25.f{hrs:03d}')
+            outfile = os.path.join(outdir, f'gencastge{gefs_member:02d}.t{cycle:02d}z.pgrb2.0p25.f{hrs:03d}')
             print(outfile)
 
             for cube in cubes:
@@ -108,8 +108,8 @@ class Netcdf2Grib:
                 cube.remove_coord('time')
                 cube.add_dim_coord(new_time_coord, time_coord_dim)
 
-                hour_6 = iris.Constraint(time=iris.time.PartialDateTime(month=date.month, day=date.day, hour=date.hour))
-                cube_slice = cube.extract(hour_6)
+                hour_12 = iris.Constraint(time=iris.time.PartialDateTime(month=date.month, day=date.day, hour=date.hour))
+                cube_slice = cube.extract(hour_12)
                 cube_slice.coord('latitude').coord_system = iris.coord_systems.GeogCS(4326)
                 cube_slice.coord('longitude').coord_system = iris.coord_systems.GeogCS(4326)
 
@@ -126,35 +126,18 @@ class Netcdf2Grib:
                     cube_slice.standard_name = self.ATTR_MAPS[var_name][1]
                     cube_slice.units = self.ATTR_MAPS[var_name][2]
 
-                    if var_name not in ['mean_sea_level_pressure', 'total_precipitation_6hr', 'total_precipitation']:
+                    if var_name not in ['mean_sea_level_pressure', 'total_precipitation_12hr', 'total_precipitation']:
                         cube_slice.add_aux_coord(iris.coords.DimCoord(self.ATTR_MAPS[var_name][0], standard_name='height', units='m'))
                         iris.save(cube_slice, outfile, saver='grib2', append=True)
-                    elif var_name == 'total_precipitation_6hr':
-                        iris_grib.save_messages(self.tweaked_messages(cube_slice, f'{hrs-6}-{hrs}'), outfile, append=True)
+                    elif var_name == 'total_precipitation_12hr':
+                        iris_grib.save_messages(self.tweaked_messages(cube_slice, f'{hrs-12}-{hrs}'), outfile, append=True)
                     elif var_name == 'total_precipitation':
                         iris_grib.save_messages(self.tweaked_messages(cube_slice, f'0-{hrs}'), outfile, append=True)
                     elif var_name == 'mean_sea_level_pressure':
                         cube_slice.add_aux_coord(iris.coords.DimCoord(self.ATTR_MAPS[var_name][0], standard_name='altitude', units='m'))
-                        iris_grib.save_messages(self.tweaked_messages(cube_slice, f'{hrs-6}-{hrs}'), outfile, append=True)
+                        iris_grib.save_messages(self.tweaked_messages(cube_slice, f'{hrs-12}-{hrs}'), outfile, append=True)
 
         # Remove intermediate netCDF file
         if os.path.isfile(filename):
             print(f'Deleting intermediate nc file {filename}: ')
             os.remove(filename)
-
-        # subset grib2 files
-        def subset_grib2(indir=None):
-            files = glob.glob(f'{indir}/graphcastgfs.*')
-            files.sort()
-        
-            outdir = os.path.join(indir, 'north_america')
-            os.makedirs(outdir, exist_ok=True)
-            
-            lonMin, lonMax, latMin, latMax = 61.0, 299.0, -37.0, 37.0 
-            for grbfile in files:
-                outfile = f"{outdir}/{grbfile.split('/')[-1]}"
-                command = ['wgrib2', grbfile, '-small_grib', f'{lonMin}:{lonMax}', f'{latMin}:{latMax}', outfile]
-                subprocess.run(command, check=True)
-                
-        
-        # subset_grib2(outdir)
