+++ readme.md
@@ -51,8 +51,15 @@ This will install the packages and most of their dependencies.
 
 
 ## Usage
+### Input data
+```bash
+python gdas_utils.py YYYYMMDDHH -l 13 -m wgrib2(or pygrib) -s s3 -l /path/to/output -d /path/to/download -k no
+````
 
-
+### Run the model
+```bash
+python run_gencast.py -i /path/to/inputfile -w /path/to/model/ -l lead_time(steps) -m num_of_ensemble_members -o /path/to/output -p num_of_pls -u yes(no) -k yes(no)
+````
 
 ## Output
